Downloads/ai-legal-demand-starter/ai-legal-demand/requirements.txt
python-dotenv
pdfplumber
PyPDF2
openai>=1.0.0
chromadb
rank-bm25
tiktoken
python-docx
docx2pdf ; platform_system == "Windows"
pytest
fastapi
uvicorn
